Mulighet for mer nøyaktig validering av endringer fra punsj
